2004-07-22  Jay Rossiter <cryptographite@users.sf.net>  1.4.5-cvs-b57
        * Change: Better handling of the onsubmit/double-clicking issue
          utilizing an 'action' input which is changed depending on
          which button was clicked.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@6884 57fcd75e-5312-0410-8df3-f5eb6fbb1595

+++ delete_album.php
@@ -33,7 +33,7 @@ if (!$gallery->user->canDeleteAlbum($gallery->album)) {
 doctype();
 echo "\n<html>";
 
-if (!empty($delete)) {
+if (!empty($action) && $action == 'delete') {
 	if ($guid == $gallery->album->fields['guid']) {
 		$gallery->album->delete();
 	}
@@ -59,7 +59,8 @@ if ($gallery->album) {
 <?php echo makeFormIntro("delete_album.php", array('name' => 'deletealbum_form', 'onsubmit' => 'deletealbum_form.delete.disabled
 = true;')); ?>
 <input type="hidden" name="guid" value="<?php echo $gallery->album->fields['guid']; ?>">
-<input type="submit" name="delete" value="<?php echo _("Delete") ?>">
+<input type="hidden" name="action" value="">
+<input type="submit" name="delete" value="<?php echo _("Delete") ?>" onclick="deletealbum_form.action.value='delete'">
 <input type="button" name="cancel" value="<?php echo _("Cancel") ?>" onclick='parent.close()'>
 </form>
 <p>
